+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Breaking change note:
 resource definition from the Go SDK. Following this change, use the Provider resource at
 `github.com/pulumi/pulumi-kubernetes/sdk/v3/go/kubernetes` instead.
 
-- Helm Release: Make RepositoryOpts optional  (https://github.com/pulumi/pulumi-kubernetes/pull/1806)
+- Helm Release: Make RepositoryOpts optional (https://github.com/pulumi/pulumi-kubernetes/pull/1806)
+- Helm Release: Support local charts (https://github.com/pulumi/pulumi-kubernetes/pull/1809)
 - Update pulumi dependencies v3.19.0 (https://github.com/pulumi/pulumi-kubernetes/pull/1816)
 - [go/sdk] Remove deprecated providers/Provider resource (https://github.com/pulumi/pulumi-kubernetes/pull/1817)